Use the production tegola image
The development image is designed to be rebuilt any time the source
changes, and something in the build steps change the directory
contents thereby causing a full rebuild every run.  Using the
production image will be more efficient as long as we aren't changing
tegola source, because it will never need a rebuild.